# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/books/{id}":{"get":{"tags":["book-controller"],"operationId":"getBook","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookDTO"}}}}}},"put":{"tags":["book-controller"],"operationId":"updateBook","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookDTO"}}}}}},"delete":{"tags":["book-controller"],"operationId":"deleteBook","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/authors/{id}":{"get":{"tags":["author-controller"],"operationId":"getAuthor","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthorDTO"}}}}}},"put":{"tags":["author-controller"],"operationId":"updateAuthor","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthorDTO"}}}}}},"delete":{"tags":["author-controller"],"operationId":"deleteAuthor","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/books":{"get":{"tags":["book-controller"],"operationId":"getAllBooks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookDTO"}}}}}}},"post":{"tags":["book-controller"],"operationId":"createBook","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Book"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookDTO"}}}}}}},"/authors":{"get":{"tags":["author-controller"],"operationId":"getAllAuthors","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuthorDTO"}}}}}}},"post":{"tags":["author-controller"],"operationId":"createAuthor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Author"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthorDTO"}}}}}}}},"components":{"schemas":{"Author":{"required":["email","name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"maxLength":50,"minLength":2,"type":"string"},"email":{"type":"string"},"books":{"type":"array","items":{"$ref":"#/components/schemas/Book"}}}},"Book":{"required":["author","title"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"maxLength":100,"minLength":1,"type":"string"},"author":{"$ref":"#/components/schemas/Author"}}},"AuthorDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"books":{"type":"array","items":{"$ref":"#/components/schemas/BookDTO"}}}},"BookDTO":{"type":"object","properties":{"title":{"type":"string"},"author":{"$ref":"#/components/schemas/AuthorDTO"}}}}}}