# src/main/java/com/asm/ws/rest4j/dto/AuthorDTO.java
package com.asm.ws.rest4j.dto;

import java.util.List;

/**
 * @author ashut
 * @since 31-08-2024
 */

public class AuthorDTO {
    private Long id;
    private String name;
    private String email;
    private List<BookDTO> books;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public List<BookDTO> getBooks() {
        return books;
    }

    public void setBooks(List<BookDTO> books) {
        this.books = books;
    }
}